+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Method dumpPrint() to print HTML code in the export form
 - Method editRowPrint() receives the query which got the edited row and the time it took
 - Plugin config: Save the settings by POST so that the CSRF token can be verified
+- Plugin sql-log: Write the log to the temp directory by default
 - Driver plugins: Driver::select() assigns the executed query to $this->query
 - Driver plugins: Bundle the syntax highlighter in the plugin file, adminer-plugins/jush-&lt;driver>.js is not loaded anymore
 - Driver plugins: Method jushAutocomplete() to disable or replace the autocomplete of the query &lt;textarea>